--- src/App.tsx
@@ -1,13 +1,0 @@
-
-import Form from "./Form";
-
-function App(): JSX.Element {  
-
-  return(
-    <div>
-      <Form/>
-    </div>
-  );
-}
-
-export default App;
