+++ src/components/App.tsx
@@ -1,0 +1,11 @@
+import Form from './Form';
+
+function App(): JSX.Element {
+  return (
+    <div>
+      <Form />
+    </div>
+  );
+}
+
+export default App;
