+++ src/helpers/handleLoginClick.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export const handleLoginClick = (
+  setIsSignupPage: Dispatch<SetStateAction<boolean>>
+) => {
+  setIsSignupPage(false);
+};
