+++ src/helpers/handleSignupClick.ts
@@ -1,0 +1,7 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export const handleSignupClick = (
+  setIsSignupPage: Dispatch<SetStateAction<boolean>>
+) => {
+  setIsSignupPage(true);
+};
