+++ tsconfig.app.json
@@ -3,7 +3,7 @@
     "composite": true,
     "target": "es6",
     "useDefineForClassFields": true,
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "lib": ["ES2020", "es2015", "DOM", "DOM.Iterable"],
     "module": "esnext",
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -23,6 +23,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules"]
 }
