+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
   "files": [
-    "./src/App.tsx",
-    "./src/Form.tsx",
-    "./src/main.tsx",
+    "./src/components/App.tsx",
+    "src/components/Form.tsx",
+    "src/components/main.tsx",
     "./src/vite-env.d.ts"
   ],
   "references": [{ "path": "./tsconfig.app.json" }]
