+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/form.tsx","./src/main.tsx","./src/vite-env.d.ts"],"errors":true,"version":"5.6.3"}
+{"root":["./src/components/app.tsx","./src/components/form.tsx","./src/components/main.tsx","./src/vite-env.d.ts"],"version":"5.6.3"}
